3.1.17 SFA 16/10/2021

+++ app/src/main/java/com/hap/checkinproc/SFA_Activity/Invoice_Category_Select.java
@@ -129,7 +129,7 @@ public class Invoice_Category_Select extends AppCompatActivity implements View.O
 
     String orderId = "";
     private LinearLayout rlAddProduct, rlCredit, rlCash;
-    private int outstandAmt;
+    private double outstandAmt;
     private double payAmt;
 
     final Handler handler = new Handler();
